Moved 'remove_low_intensity' to 'method'.
